FUSETOOLS-3057 - extract runnable to specific class and reduce complexity
